Inline ansi modules
